lara/col/land: move wading to land module
Wading collision control has more in common with land routines,
so this moves it back to that module. This will allow further misc
functions to become private there.
