+++ src/libtrx/game/lara/col/land.c
@@ -1,4 +1,5 @@
 #include "config.h"
+#include "game/input.h"
 #include "game/lara.h"
 #include "game/lara/util.h"
 #include "game/rooms.h"
@@ -23,10 +24,19 @@
 #define M_LF_RUN_R_FOOT_GROUND 14
 #define M_LF_RUN_R_END 21
 
+#define M_LF_WADE_L_START 0
+#define M_LF_WADE_L_END 9
+#define M_LF_WADE_R_START 10
+#define M_LF_WADE_R_END 21
+#define M_LF_WADE_STEP_L_START 3
+#define M_LF_WADE_STEP_L_END 14
+
 static bool M_FixDescendingGlitch(void);
 static bool M_FixStepGlitch(void);
 static bool M_TestWall(
     const ITEM *item, int32_t front, int32_t right, int32_t down);
+static bool M_Fallen(ITEM *item, const COLL_INFO *coll);
+static void M_CollideStop(ITEM *item, const COLL_INFO *coll);
 
 static void M_Default(ITEM *item, COLL_INFO *coll);
 static void M_Walk(ITEM *item, COLL_INFO *coll);
@@ -41,6 +51,7 @@ static void M_Splat(ITEM *item, COLL_INFO *coll);
 static void M_Slide(ITEM *item, COLL_INFO *coll);
 static void M_Roll(ITEM *item, COLL_INFO *coll);
 static void M_RollContinue(ITEM *item, COLL_INFO *coll);
+static void M_Wade(ITEM *item, COLL_INFO *coll);
 
 static bool M_FixDescendingGlitch(void)
 {
@@ -115,6 +126,57 @@ static bool M_TestWall(
     return true;
 }
 
+static bool M_Fallen(ITEM *const item, const COLL_INFO *const coll)
+{
+    const LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (coll->side_mid.floor <= STEPUP_HEIGHT
+        || lara->water_status == LWS_WADE) {
+        return false;
+    }
+    item->current_anim_state = LS_JUMP_FORWARD;
+    item->goal_anim_state = LS_JUMP_FORWARD;
+    Item_SwitchToAnim(item, LA_FALL_START, 0);
+    item->gravity = true;
+    item->fall_speed = 0;
+    return true;
+}
+
+static void M_CollideStop(ITEM *const item, const COLL_INFO *const coll)
+{
+#if TR_VERSION == 1
+    // TODO: this routine gives smoother recovery after splatting against a wall
+    // - offer it fully in TR1 as its only scope is currently for wading.
+    const LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (lara->water_status != LWS_WADE) {
+        Item_SwitchToAnim(item, LA_STAND_STILL, 0);
+        return;
+    }
+#endif
+
+    switch (coll->old_anim_state) {
+    case LS_STOP:
+    case LS_TURN_RIGHT:
+    case LS_TURN_LEFT:
+    case LS_FAST_TURN:
+        item->current_anim_state = coll->old_anim_state;
+        item->anim_num = coll->old_anim_num;
+        item->frame_num = coll->old_frame_num;
+        if (g_Input.left) {
+            item->goal_anim_state = LS_TURN_LEFT;
+        } else if (g_Input.right) {
+            item->goal_anim_state = LS_TURN_RIGHT;
+        } else {
+            item->goal_anim_state = LS_STOP;
+        }
+        Lara_Animate(item);
+        break;
+
+    default:
+        Item_SwitchToAnim(item, LA_STAND_STILL, 0);
+        break;
+    }
+}
+
 static void M_Default(ITEM *const item, COLL_INFO *const coll)
 {
     LARA_INFO *const lara = Lara_GetLaraInfo();
@@ -151,11 +213,11 @@ static void M_Walk(ITEM *const item, COLL_INFO *const coll)
                     item, M_LF_WALK_STEP_L_2_START, M_LF_WALK_STEP_L_2_END))) {
             Item_SwitchToAnim(item, LA_WALK_STOP_LEFT, 0);
         } else {
-            Lara_Col_Stop(item, coll);
+            M_CollideStop(item, coll);
         }
     }
 
-    if (Lara_Col_Fallen(item, coll)) {
+    if (M_Fallen(item, coll)) {
         return;
     }
 
@@ -209,10 +271,10 @@ static void M_WalkBack(ITEM *const item, COLL_INFO *const coll)
     }
 
     if (Lara_DeflectEdge(item, coll)) {
-        Lara_Col_Stop(item, coll);
+        M_CollideStop(item, coll);
     }
 
-    if (M_FixDescendingGlitch() && Lara_Col_Fallen(item, coll)) {
+    if (M_FixDescendingGlitch() && M_Fallen(item, coll)) {
         return;
     }
 
@@ -258,10 +320,10 @@ static void M_SideStep(ITEM *const item, COLL_INFO *const coll)
     }
 
     if (Lara_DeflectEdge(item, coll)) {
-        Lara_Col_Stop(item, coll);
+        M_CollideStop(item, coll);
     }
 
-    if (M_FixDescendingGlitch() && Lara_Col_Fallen(item, coll)) {
+    if (M_FixDescendingGlitch() && M_Fallen(item, coll)) {
         return;
     }
 
@@ -307,10 +369,10 @@ static void M_Run(ITEM *const item, COLL_INFO *const coll)
                 return;
             }
         }
-        Lara_Col_Stop(item, coll);
+        M_CollideStop(item, coll);
     }
 
-    if (Lara_Col_Fallen(item, coll)) {
+    if (M_Fallen(item, coll)) {
         return;
     }
 
@@ -344,7 +406,7 @@ static void M_Stop(ITEM *const item, COLL_INFO *const coll)
     M_Default(item, coll);
 
     if (Lara_HitCeiling(item, coll)
-        || (M_FixDescendingGlitch() && Lara_Col_Fallen(item, coll))
+        || (M_FixDescendingGlitch() && M_Fallen(item, coll))
         || Lara_TestSlide(item, coll)) {
         return;
     }
@@ -381,7 +443,7 @@ static void M_FastBack(ITEM *const item, COLL_INFO *const coll)
 
     if (coll->side_mid.floor <= 200) {
         if (Lara_DeflectEdge(item, coll)) {
-            Lara_Col_Stop(item, coll);
+            M_CollideStop(item, coll);
         }
         item->pos.y += coll->side_mid.floor;
     } else {
@@ -480,7 +542,7 @@ static void M_Roll(ITEM *const item, COLL_INFO *const coll)
         return;
     }
 #else
-    if (Lara_Col_Fallen(item, coll)) {
+    if (M_Fallen(item, coll)) {
         return;
     }
 #endif
@@ -515,6 +577,60 @@ static void M_RollContinue(ITEM *const item, COLL_INFO *const coll)
         Lara_ShiftCol(coll);
         item->pos.y += coll->side_mid.floor;
     }
+}
+
+static void M_Wade(ITEM *const item, COLL_INFO *const coll)
+{
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+    coll->slopes_are_walls = 1;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = 0;
+
+    Lara_GetCollisionInfo(item, coll);
+    if (Lara_HitCeiling(item, coll) || Lara_TestVault(item, coll)) {
+        return;
+    }
+
+    if (Lara_DeflectEdge(item, coll)) {
+        item->rot.z = 0;
+        if (coll->side_front.type == HT_WALL
+            && coll->side_front.floor < -STEP_L * 5 / 2
+            && coll->old_anim_state == LS_WADE
+            && Item_TestAnimEqual(item, LA_WADE)) {
+            item->current_anim_state = LS_SPLAT;
+            if (Item_TestFrameRange(item, M_LF_WADE_L_START, M_LF_WADE_L_END)) {
+                Item_SwitchToAnim(item, LA_WALL_SMASH_LEFT, 0);
+                return;
+            }
+            if (Item_TestFrameRange(item, M_LF_WADE_R_START, M_LF_WADE_R_END)) {
+                Item_SwitchToAnim(item, LA_WALL_SMASH_RIGHT, 0);
+                return;
+            }
+        }
+        M_CollideStop(item, coll);
+    }
+
+    if (M_Fallen(item, coll)) {
+        return;
+    }
+
+    if (coll->side_mid.floor >= -STEPUP_HEIGHT
+        && coll->side_mid.floor < -STEP_L / 2) {
+        if (Item_TestFrameRange(
+                item, M_LF_WADE_STEP_L_START, M_LF_WADE_STEP_L_END)) {
+            Item_SwitchToAnim(item, LA_RUN_UP_STEP_LEFT, 0);
+        } else {
+            Item_SwitchToAnim(item, LA_RUN_UP_STEP_RIGHT, 0);
+        }
+    }
+
+    if (Lara_TestSlide(item, coll)) {
+        return;
+    }
+
+    item->pos.y += MIN(coll->side_mid.floor, 50);
 }
 
 // clang-format off
@@ -554,4 +670,5 @@ REGISTER_LARA_COL(LS_SLIDE,        M_Slide)
 REGISTER_LARA_COL(LS_SLIDE_BACK,   M_Slide)
 REGISTER_LARA_COL(LS_ROLL,         M_Roll)
 REGISTER_LARA_COL(LS_ROLL_CONT,    M_RollContinue)
+REGISTER_LARA_COL(LS_WADE,         M_Wade)
 // clang-format on
