+++ src/libtrx/include/libtrx/game/lara/col.h
@@ -6,5 +6,3 @@
 void Lara_Col_Register(
     LARA_STATE state, void (*handle_func)(ITEM *item, COLL_INFO *coll));
 void Lara_Col_Update(ITEM *item, COLL_INFO *coll);
-bool Lara_Col_Fallen(ITEM *item, const COLL_INFO *coll);
-void Lara_Col_Stop(ITEM *item, const COLL_INFO *coll);
